Update MagmaFunctions.m
Updated Hankel and HankelKer. Now you can ask HankelKer with efficient := true to use smaller Hankels when possible.
The IsPower function has been removed (it would have been more efficient, but depends on the defining field. Hence, it may lead to errors in rare cases, which we prefer to avoid it).